+++ apps/api-docs/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   `replace_uuid` parameter for `postUpload` operation.
+-   `existing_uuid` parameter for `postUpload` operation.
 
 ### Changed